Refactor navigation to Header menu and move Settings to a dedicated view
- Replaced Sidebar navigation with a Header component containing a single "Menu" icon.
- Implemented a vertical dropdown menu triggered by hover/click on the Menu icon.
- The menu includes a fixed "Views" submenu and dynamic context-aware submenus.
- Converted the Settings modal into a full-page `SettingsView`.
- Refactored `appState.svelte.js` to manage dynamic menu actions (`menuActions`).
- Updated all major views (`DocumentsPage`, `PartsView`, `GncView`, `JobView`, `StockView`, `JournalView`, `DashboardView`) to register their specific actions (e.g., "New Document", "Save", "Refresh") into the global menu on mount.
- Removed obsolete `Sidebar.svelte` and `SettingsModal.svelte`.
